IMPR: i2c refactor finished
